Move RecordEndpoints class to an Endpoints folder
